style: move the upload form into new directory for upload page
